@W-17589798@ Bump develop to v3.10.0-dev (#2278)
* Update changelog files

* Bump to the preview versions

* Bump up preview versions

* @W-17589798@ Release preview version 3.9.0-preview.2 (#2264)

* Update default.js.hbs and ssr.js.hbs templates to support new login flows (#2263)

* update bootstrap default.js.hbs to include login options
* update ssr.js templates

* Revert/page meta data tags 3.9.x (#2259)

* Revert page-meta-data changes

* Update CHANGELOG.md

* Fix changelog

---------


* Bump to version 3.9.0-preview.2

* Fix lint errors in default.js templates (#2265)

* fix lint errors

* rerun ci jobs

---------


* @W-17589798@ Release preview version 3.9.0-preview.3 (#2268)

* Update default.js.hbs and ssr.js.hbs templates to support new login flows (#2263)

* update bootstrap default.js.hbs to include login options
* update ssr.js templates

* Revert/page meta data tags 3.9.x (#2259)

* Revert page-meta-data changes

* Update CHANGELOG.md

* Fix changelog

---------


* Bump to version 3.9.0-preview.2

* Fix lint errors in default.js templates (#2265)

* fix lint errors

* rerun ci jobs

* Update app config templates (#2267)

* lint default.js in templates folder (#2266)


* Bump to version 3.9.0-preview.3

* Update CHANGELOG.md

---------


* @W-17589798@ Release preview version 3.9.0-preview.4 (#2272)

* Update default.js.hbs and ssr.js.hbs templates to support new login flows (#2263)

* update bootstrap default.js.hbs to include login options
* update ssr.js templates

* Revert/page meta data tags 3.9.x (#2259)

* Revert page-meta-data changes

* Update CHANGELOG.md

* Fix changelog

---------


* Bump to version 3.9.0-preview.2

* Fix lint errors in default.js templates (#2265)

* fix lint errors

* rerun ci jobs

* Update app config templates (#2267)

* lint default.js in templates folder (#2266)


* Bump to version 3.9.0-preview.3

* Update CHANGELOG.md

* Fix import error template retail react app (#2270)

* revert babel.config.js workaround

* move everything out of utils into ssr.js or default.js

* update default.js templates

* update ssr.js templates

* rerun ci jobs

* rerun ci jobs

* Bump versions

---------


* Update CHANGELOG.md

---------


* @W-17589798@ Release preview version 3.9.0-preview.5 (#2274)

* Update default.js.hbs and ssr.js.hbs templates to support new login flows (#2263)

* update bootstrap default.js.hbs to include login options
* update ssr.js templates

* Revert/page meta data tags 3.9.x (#2259)

* Revert page-meta-data changes

* Update CHANGELOG.md

* Fix changelog

---------


* Bump to version 3.9.0-preview.2

* Fix lint errors in default.js templates (#2265)

* fix lint errors

* rerun ci jobs

* Update app config templates (#2267)

* lint default.js in templates folder (#2266)


* Bump to version 3.9.0-preview.3

* Update CHANGELOG.md

* Fix import error template retail react app (#2270)

* revert babel.config.js workaround

* move everything out of utils into ssr.js or default.js

* update default.js templates

* update ssr.js templates

* rerun ci jobs

* rerun ci jobs

* Bump versions

---------


* Update CHANGELOG.md

* Add missing `_app-config` template changes (#2273)

* @W-17589798@ Release preview version 3.9.0-preview.4 (#2272)

* Update default.js.hbs and ssr.js.hbs templates to support new login flows (#2263)

* update bootstrap default.js.hbs to include login options
* update ssr.js templates

* Revert/page meta data tags 3.9.x (#2259)

* Revert page-meta-data changes

* Update CHANGELOG.md

* Fix changelog

---------


* Bump to version 3.9.0-preview.2

* Fix lint errors in default.js templates (#2265)

* fix lint errors

* rerun ci jobs

* Update app config templates (#2267)

* lint default.js in templates folder (#2266)


* Bump to version 3.9.0-preview.3

* Update CHANGELOG.md

* Fix import error template retail react app (#2270)

* revert babel.config.js workaround

* move everything out of utils into ssr.js or default.js

* update default.js templates

* update ssr.js templates

* rerun ci jobs

* rerun ci jobs

* Bump versions

---------


* Update CHANGELOG.md

---------


* add missing app config template changes

* update both app config templates

---------


* Bump version to 3.9.0-preview.5

* Update CHANGELOG.md

---------


* @W-17589798@ PWA Kit v3.9.0, commerce-sdk-react 3.2.0, retail-react-app 6.0.0 Release (#2277)

* Update default.js.hbs and ssr.js.hbs templates to support new login flows (#2263)

* update bootstrap default.js.hbs to include login options
* update ssr.js templates

* Revert/page meta data tags 3.9.x (#2259)

* Revert page-meta-data changes

* Update CHANGELOG.md

* Fix changelog

---------


* Bump to version 3.9.0-preview.2

* Fix lint errors in default.js templates (#2265)

* fix lint errors

* rerun ci jobs

* Update app config templates (#2267)

* lint default.js in templates folder (#2266)


* Bump to version 3.9.0-preview.3

* Update CHANGELOG.md

* Fix import error template retail react app (#2270)

* revert babel.config.js workaround

* move everything out of utils into ssr.js or default.js

* update default.js templates

* update ssr.js templates

* rerun ci jobs

* rerun ci jobs

* Bump versions

---------


* Update CHANGELOG.md

* Add missing `_app-config` template changes (#2273)

* @W-17589798@ Release preview version 3.9.0-preview.4 (#2272)

* Update default.js.hbs and ssr.js.hbs templates to support new login flows (#2263)

* update bootstrap default.js.hbs to include login options
* update ssr.js templates

* Revert/page meta data tags 3.9.x (#2259)

* Revert page-meta-data changes

* Update CHANGELOG.md

* Fix changelog

---------


* Bump to version 3.9.0-preview.2

* Fix lint errors in default.js templates (#2265)

* fix lint errors

* rerun ci jobs

* Update app config templates (#2267)

* lint default.js in templates folder (#2266)


* Bump to version 3.9.0-preview.3

* Update CHANGELOG.md

* Fix import error template retail react app (#2270)

* revert babel.config.js workaround

* move everything out of utils into ssr.js or default.js

* update default.js templates

* update ssr.js templates

* rerun ci jobs

* rerun ci jobs

* Bump versions

---------


* Update CHANGELOG.md

---------


* add missing app config template changes

* update both app config templates

---------


* Bump version to 3.9.0-preview.5

* Update CHANGELOG.md

* Bump versions

* Update changelogs

* Update CHANGELOG.md

* Update peer dep in package-lock

---------


* Bump version to 3.10.0-dev

---------
