+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.9.0-dev",
+  "version": "3.10.0-dev",
   "packages": [
     "packages/*"
   ]
