+++ packages/internal-lib-build/package.json
@@ -1,6 +1,6 @@
 {
   "name": "internal-lib-build",
-  "version": "3.9.0-dev",
+  "version": "3.10.0-dev",
   "private": true,
   "description": "Build tools for *libraries* in the monorepo",
   "bugs": {
@@ -60,7 +60,7 @@
     "shelljs": "^0.8.5"
   },
   "devDependencies": {
-    "@salesforce/pwa-kit-dev": "3.9.0-dev",
+    "@salesforce/pwa-kit-dev": "3.10.0-dev",
     "npm-packlist": "^4.0.0",
     "typescript": "4.9.5"
   },
