+++ packages/pwa-kit-create-app/assets/templates/@salesforce/retail-react-app/config/default.js.hbs
@@ -26,8 +26,10 @@ module.exports = {
                 // The callback URI, which can be an absolute URL (including third-party URIs) or a relative path set up by the developer.
                 // Required in 'callback' mode; if missing, passwordless login defaults to 'sms' mode, which requires Marketing Cloud configuration.
                 // If the env var `PASSWORDLESS_LOGIN_CALLBACK_URI` is set, it will override the config value.
-                callbackURI: 
-                    process.env.PASSWORDLESS_LOGIN_CALLBACK_URI || '/passwordless-login-callback'
+                callbackURI:
+                    process.env.PASSWORDLESS_LOGIN_CALLBACK_URI || '/passwordless-login-callback',
+                // The landing path for passwordless login
+                landingPath: '/passwordless-login-landing'
             },
             social: {
                 // Enables or disables social login for the site. Defaults to: false
@@ -35,7 +37,7 @@ module.exports = {
                 // The third-party identity providers supported by your app. The PWA Kit supports Google and Apple by default.
                 // Additional IDPs will also need to be added to the IDP_CONFIG in the SocialLogin component.
                 idps: ['google', 'apple'],
-                // The redirect URI used after a successful social login authentication. 
+                // The redirect URI used after a successful social login authentication.
                 // This should be a relative path set up by the developer.
                 // If the env var `SOCIAL_LOGIN_REDIRECT_URI` is set, it will override the config value.
                 redirectURI: process.env.SOCIAL_LOGIN_REDIRECT_URI || '/social-callback'
@@ -43,7 +45,9 @@ module.exports = {
             resetPassword: {
                 // The callback URI, which can be an absolute URL (including third-party URIs) or a relative path set up by the developer.
                 // If the env var `RESET_PASSWORD_CALLBACK_URI` is set, it will override the config value.
-                callbackURI: process.env.RESET_PASSWORD_CALLBACK_URI || '/reset-password-callback'
+                callbackURI: process.env.RESET_PASSWORD_CALLBACK_URI || '/reset-password-callback',
+                // The landing path for reset password
+                landingPath: '/reset-password-landing'
             }
         },
         // The default site for your app. This value will be used when a siteRef could not be determined from the url
