+++ packages/pwa-kit-dev/CHANGELOG.md
@@ -1,7 +1,12 @@
-## v3.9.0-dev (Oct 29, 2024)
-- Support Node 22 [#2218](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2218)
+## v3.10.0-dev (Feb 18, 2025)
+
 - Support source map for both client and server on MRT [#240](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2240)
-- 
+
+## v3.9.0 (Feb 18, 2025)
+
+- Support Node 22 [#2218](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2218) 
+- Set minimum target node version [#2241](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2241)
+
 ## v3.8.0 (Oct 28, 2024)
 
 ## v3.7.0 (Aug 07, 2024)
