+++ packages/pwa-kit-runtime/CHANGELOG.md
@@ -1,4 +1,6 @@
-## v3.9.0-dev (Oct 29, 2024)
+## v3.10.0-dev (Feb 18, 2025)
+
+## v3.9.0 (Feb 18, 2025)
 - Fix stale service worker file that could cause requests to still use old Content-Security-Policy [#2191](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2191)
 - Support Node 22 [#2218](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2218)
 - Support Social Login, Passwordless Login, and Password Reset: update the default value for `applySLASPrivateClientToEndpoints` option [#2250](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2250)
