+++ packages/pwa-kit-runtime/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@salesforce/pwa-kit-runtime",
-  "version": "3.9.0-dev",
+  "version": "3.10.0-dev",
   "description": "The PWAKit Runtime",
   "homepage": "https://github.com/SalesforceCommerceCloud/pwa-kit/tree/develop/packages/pwa-kit-runtime#readme",
   "bugs": {
@@ -46,11 +46,11 @@
   },
   "devDependencies": {
     "@loadable/component": "^5.15.3",
-    "@salesforce/pwa-kit-dev": "3.9.0-dev",
+    "@salesforce/pwa-kit-dev": "3.10.0-dev",
     "@serverless/event-mocks": "^1.1.1",
     "aws-lambda-mock-context": "^3.2.1",
     "fs-extra": "^11.1.1",
-    "internal-lib-build": "3.9.0-dev",
+    "internal-lib-build": "3.10.0-dev",
     "nock": "^13.3.0",
     "nodemon": "^2.0.22",
     "sinon": "^13.0.2",
@@ -58,7 +58,7 @@
     "supertest": "^4.0.2"
   },
   "peerDependencies": {
-    "@salesforce/pwa-kit-dev": "3.9.0-dev"
+    "@salesforce/pwa-kit-dev": "3.10.0-dev"
   },
   "peerDependenciesMeta": {
     "@salesforce/pwa-kit-dev": {
