+++ packages/template-mrt-reference-app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "template-mrt-reference-app",
-  "version": "3.9.0-dev",
+  "version": "3.10.0-dev",
   "private": true,
   "license": "See license in LICENSE",
   "scripts": {
@@ -16,8 +16,8 @@
   },
   "devDependencies": {
     "@loadable/component": "^5.15.3",
-    "@salesforce/pwa-kit-dev": "3.9.0-dev",
-    "@salesforce/pwa-kit-runtime": "3.9.0-dev",
+    "@salesforce/pwa-kit-dev": "3.10.0-dev",
+    "@salesforce/pwa-kit-runtime": "3.10.0-dev",
     "@smithy/smithy-client": "^2.1.15",
     "aws-sdk-client-mock": "^3.0.0",
     "cross-fetch": "^3.1.4",
