+++ packages/template-retail-react-app/app/pages/product-detail/index.jsx
@@ -99,8 +99,7 @@ const ProductDetail = () => {
                     'prices',
                     'variations',
                     'set_products',
-                    'bundled_products',
-                    'page_meta_tags'
+                    'bundled_products'
                 ],
                 allImages: true
             }
@@ -456,15 +455,7 @@ const ProductDetail = () => {
         >
             <Helmet>
                 <title>{product?.pageTitle}</title>
-                {product?.pageMetaTags?.length > 0 &&
-                    product.pageMetaTags.map(({id, value}) => (
-                        <meta name={id} content={value} key={id} />
-                    ))}
-                {/* Fallback for description if not included in pageMetaTags */}
-                {!product?.pageMetaTags?.some((tag) => tag.id === 'description') &&
-                    product?.pageDescription && (
-                        <meta name="description" content={product.pageDescription} />
-                    )}
+                <meta name="description" content={product?.pageDescription} />
             </Helmet>
 
             <Stack spacing={16}>
