+++ packages/template-retail-react-app/app/pages/product-list/index.jsx
@@ -159,14 +159,7 @@ const ProductList = (props) => {
                 perPricebook: true,
                 allVariationProperties: true,
                 allImages: true,
-                expand: [
-                    'promotions',
-                    'variations',
-                    'prices',
-                    'images',
-                    'page_meta_tags',
-                    'custom_properties'
-                ],
+                expand: ['promotions', 'variations', 'prices', 'images', 'custom_properties'],
                 refine: _refine
             }
         },
@@ -418,9 +411,6 @@ const ProductList = (props) => {
                 <title>{category?.pageTitle ?? searchQuery}</title>
                 <meta name="description" content={category?.pageDescription ?? searchQuery} />
                 <meta name="keywords" content={category?.pageKeywords} />
-                {productSearchResult?.pageMetaTags?.map(({id, value}) => {
-                    return <meta name={id} content={value} key={id} />
-                })}
             </Helmet>
             {showNoResults ? (
                 <EmptySearchResults searchQuery={searchQuery} category={category} />
