+++ packages/test-commerce-sdk-react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "test-commerce-sdk-react",
-  "version": "3.9.0-dev",
+  "version": "3.10.0-dev",
   "private": true,
   "scripts": {
     "build": "pwa-kit-dev build",
@@ -18,17 +18,17 @@
   },
   "devDependencies": {
     "@loadable/component": "^5.15.3",
-    "@salesforce/commerce-sdk-react": "3.2.0-dev",
-    "@salesforce/pwa-kit-dev": "3.9.0-dev",
-    "@salesforce/pwa-kit-react-sdk": "3.9.0-dev",
-    "@salesforce/pwa-kit-runtime": "3.9.0-dev",
+    "@salesforce/commerce-sdk-react": "3.3.0-dev",
+    "@salesforce/pwa-kit-dev": "3.10.0-dev",
+    "@salesforce/pwa-kit-react-sdk": "3.10.0-dev",
+    "@salesforce/pwa-kit-runtime": "3.10.0-dev",
     "@tanstack/react-query": "^4.28.0",
     "@types/loadable__component": "~5.13.4",
     "@types/node": "~16.0.3",
     "@types/react": "~18.2.0",
     "@types/react-dom": "~18.2.1",
     "@types/react-router-dom": "~5.3.3",
-    "internal-lib-build": "3.9.0-dev",
+    "internal-lib-build": "3.10.0-dev",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
